swap: cashout strategy based on transaction costs instead of gasPrice

+++ swap/config.go
@@ -24,8 +24,8 @@ import (
 // Need experimentation to arrive to values which make sense
 const (
 	// Thresholds which trigger payment or disconnection. The unit is in honey (internal accounting unit)
-	DefaultPaymentThreshold    = int64(2441 * (RetrieveRequestPrice + ChunkDeliveryPrice)) // Payment when the disbalance is an equivalent of 10 mb of data (2441 chunks)
-	DefaultDisconnectThreshold = int64(2 * DefaultPaymentThreshold)
+	DefaultPaymentThreshold    = 2441 * (RetrieveRequestPrice + ChunkDeliveryPrice) // Payment when the disbalance is an equivalent of 10 mb of data (2441 chunks)
+	DefaultDisconnectThreshold = 2 * DefaultPaymentThreshold
 	// DefaultInitialDepositAmount is the default amount to send to the contract when initially deploying
 	// NOTE: deliberate value for now; needs experimentation
 	DefaultInitialDepositAmount = 0
