+++ swap/swap_test.go
@@ -598,7 +598,7 @@ func TestDisconnectThreshold(t *testing.T) {
 	testPeer := newDummyPeer()
 	testDeploy(context.Background(), swap)
 	swap.addPeer(testPeer.Peer, swap.owner.address, swap.GetParams().ContractAddress)
-	swap.Add(DefaultDisconnectThreshold, testPeer.Peer)
+	swap.Add(int64(DefaultDisconnectThreshold), testPeer.Peer)
 	err := swap.Add(1, testPeer.Peer)
 	if !strings.Contains(err.Error(), "disconnect threshold") {
 		t.Fatal(err)
@@ -612,13 +612,13 @@ func TestPaymentThreshold(t *testing.T) {
 	testDeploy(context.Background(), swap)
 	testPeer := newDummyPeerWithSpec(Spec)
 	swap.addPeer(testPeer.Peer, swap.owner.address, swap.GetParams().ContractAddress)
-	if err := swap.Add(-DefaultPaymentThreshold, testPeer.Peer); err != nil {
+	if err := swap.Add(-int64(DefaultPaymentThreshold), testPeer.Peer); err != nil {
 		t.Fatal()
 	}
 
 	var cheque *Cheque
 	_ = swap.store.Get(sentChequeKey(testPeer.Peer.ID()), &cheque)
-	if cheque.CumulativePayout != uint64(DefaultPaymentThreshold) {
+	if cheque.CumulativePayout != DefaultPaymentThreshold {
 		t.Fatal()
 	}
 }
@@ -765,7 +765,7 @@ func calculateExpectedBalances(swap *Swap, bookings []booking) map[enode.ID]int6
 		peerID := booking.peer.ID()
 		peerBalance := expectedBalances[peerID]
 		// balance is not expected to be affected once past the disconnect threshold
-		if peerBalance < swap.params.DisconnectThreshold {
+		if peerBalance < int64(swap.params.DisconnectThreshold) {
 			peerBalance += booking.amount
 		}
 		expectedBalances[peerID] = peerBalance
